src/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg2-binary
asyncpg
starlette
jinja2
aiofiles
titiler.core